feat(core): 修改实体类与方法中 ID 类型为 long，提升大数值兼容性
- 将 `Fid` 和 `Tid` 等字段类型从 `int` 改为 `long`，提升对大数值的支持。
- 更新相关 API 和实体类的参数类型及转换逻辑。
- 调整方法签名及调用处以适配类型调整，确保代码一致性。

+++ AioTieba4DotNet/Api/GetForum/Entities/Forum.cs
@@ -8,7 +8,7 @@ public class Forum
     /// <summary>
     ///     吧 ID
     /// </summary>
-    public int Fid { get; set; }
+    public long Fid { get; set; }
 
     /// <summary>
     ///     吧名
@@ -62,7 +62,7 @@ public class Forum
     /// <returns>吧信息实体</returns>
     internal static Forum FromTbData(IDictionary<string, object> dataMap)
     {
-        var fid = Convert.ToInt32(dataMap["id"]);
+        var fid = Convert.ToInt64(dataMap["id"]);
         var fname = dataMap["name"] as string ?? "";
         var category = dataMap["first_class"] as string ?? "";
         var subcategory = dataMap["second_class"] as string ?? "";
